:truck: moved /u/_id to the ProfileDetail
